requirement.txt
torch
torchvision
opencv-python
pandas
Pillow
matplotlib
streamlit
scikit-learn
